wifi: extend feature flag from int to long
WiFi features run out of all 32 bits of int.
Extend feature flag from int to long for more bits.

Bug: 122939035
Test: build
Test: atest frameworks/opt/net/wifi/tests/wifitests

+++ service/java/com/android/server/wifi/WifiNative.java
@@ -2616,7 +2616,7 @@ public class WifiNative {
      * @param ifaceName Name of the interface.
      * @return bitmask defined by WifiManager.WIFI_FEATURE_*
      */
-    public int getSupportedFeatureSet(@NonNull String ifaceName) {
+    public long getSupportedFeatureSet(@NonNull String ifaceName) {
         return mSupplicantStaIfaceHal.getAdvancedKeyMgmtCapabilities(ifaceName)
                 | mWifiVendorHal.getSupportedFeatureSet(ifaceName);
     }
